Refactoring set class to allow for easier extendability.   - Created a second class BlacklightOaiProvider::SolrSet that contains default implementation of sets   - BlacklightOaiProvider::Set specifies the methods that need to be implemented for SolrDocumentWrapper to work properly.
