+++ lib/blacklight_oai_provider.rb
@@ -6,6 +6,7 @@ module BlacklightOaiProvider
   autoload :Routes, 'blacklight_oai_provider/routes'
   autoload :Set, 'blacklight_oai_provider/set'
   autoload :Response, 'blacklight_oai_provider/response/list_sets'
+  autoload :SolrSet, 'blacklight_oai_provider/solr_set'
 
   require 'oai'
   require 'blacklight_oai_provider/version'
